feat: add Credentials holder for impersonated service accounts
gcloud allows developer accounts to impersonate a service account by
authenticating with the --impersonate-service-account=<service account
email> directive during login. This wraps the user credentials in
application_default_credentials.json with a representation of the
service account that can parsed and used in conjuction with
IAMCredentialsClient to sign requests with the gcloud managed
credentials for the service account.

This commit adds support for this method of URL signing by creating a
wrapper for UserRefreshCredentials that will parse the outer pieces of
the application_default_credentials.json and use that information about
the service account to make a call to the IAM Service to perform the
signing operation.

Additionally, when RequestWrapperTrait encounters
impersonate-service-account credentials for the first time it will need
to register the ImpersonatedServiceAccountCredentialsLoader factory
method with CredentialsLoader to allow for the
'impersonated_service_account' type to be handled, and create the
correct Credentials Holder object.

This can't be included in the google/auth php package since that is a
dependancy of google/cloud-iam-credentials and would create a circular
relationship.'